+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/indexes/models/SearchFieldDataType.java
@@ -76,6 +76,11 @@ public final class SearchFieldDataType extends ExpandableStringEnum<SearchFieldD
     public static final SearchFieldDataType SBYTE = fromString("Edm.SByte");
 
     /**
+     * Indicates that a field contains a 8-bit unsigned integer. This is only valid when used with Collection(Edm.Byte).
+     */
+    public static final SearchFieldDataType BYTE = fromString("Edm.Byte");
+
+    /**
      * Creates a new instance of SearchFieldDataType value.
      *
      * @deprecated Use the {@link #fromString(String)} factory method.